# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "HR Management App API Documentation",
        "contact": {
            "email": "alabaganne9@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
...
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login",
                "description": "Authenticate a user with email and password",
                "operationId": "login",
                "requestBody": {
                    "description": "User credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "johndoe@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "Password1234"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Logged in",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "type": "string",
                                            "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJodHRwOlwvXC8xMjcuMC4wLjE6ODAwMFwvYXBpXC9hdXRoXC9sb2dpbiIsImlhdCI6MTYwMTA3MTg1OSwiZXhwIjoxNjAxMTU4MjU5LCJuYmYiOjE2MDEwNzE4NTksImp0aSI6Ik9YeTdieXI3UGpkNnBaZ1kiLCJzdWIiOjMsInBydiI6IjIzYmQ1Yzg5NDlmNjAwYWRiMzllNzAxYzQwMDg3MmRiN2E1OTc2ZjcifQ.ewMSg-07JHs6YxKME0UdNMYI0CvhaE9XeWEWOSRgHRE"
                                        },
                                        "token_type": {
                                            "type": "string",
                                            "example": "bearer"
                                        },
                                        "expires_in": {
                                            "type": "integer",
                                            "example": 86400
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid credentials",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The provided credentials are incorrect."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#/components/responses/invalid-data"
                    }
                }
            }
        },
        "api/auth/me": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "User informations",
                "description": "Get current user informations",
                "operationId": "me",
                "responses": {
                    "200": {
                        "description": "User information",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/user"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/unauthenticated"
                    }
                }
            }
        },
        "/api/auth/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout",
                "description": "Logout current user",
                "operationId": "logout",
                "responses": {
                    "200": {
                        "description": "Logout response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "example": "Successfully logged out"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/unauthenticated"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "user": {
                "title": "User",
                "description": "User object",
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Ala Baganne"
                    },
                    "username": {
                        "type": "string",
                        "example": "alabaganne"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "alabaganne9@gmail.com"
                    },
                    "password": {
                        "type": "string"
                    },
                    "phone_number": {
                        "type": "string",
                        "example": "50101959"
                    },
                    "date_of_birth": {
                        "type": "string",
                        "format": "dateTime",
                        "example": "1999-12-16"
                    },
                    "address": {
                        "type": "string"
                    },
                    "gender": {
                        "type": "string",
                        "enum": [
                            "male",
                            "female"
                        ]
                    },
                    "civil_status": {
                        "enum": [
                            "single",
                            "married"
                        ]
                    },
                    "id_card_number": {
                        "type": "string"
                    },
                    "nationality": {
                        "type": "string",
                        "example": "Tunisian"
                    },
                    "university": {
                        "type": "string",
                        "example": "Higher institute of applied science and technology of sousse"
                    },
                    "history": {
                        "type": "string"
                    },
                    "experience_level": {
                        "type": "integer",
                        "maximum": 5,
                        "example": 4
                    },
                    "source": {
                        "type": "string"
                    },
                    "position": {
                        "type": "string",
                        "example": "Full stack web developer"
                    },
                    "grade": {
                        "type": "string"
                    },
                    "hiring_date": {
                        "type": "string",
                        "format": "dateTime"
                    },
                    "contract_end_date": {
                        "type": "string",
                        "format": "dateTime"
                    },
                    "type_of_contract": {
                        "type": "string",
                        "enum": [
                            "option 1",
                            "option 2",
                            "option 3"
                        ]
                    },
                    "allowed_leave_days": {
                        "type": "integer",
                        "default": 30
                    },
                    "department_id": {
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            }
        },
        "responses": {
...
            "unauthenticated": {
                "description": "Authentication is required",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "Unauthenticated."
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "invalid-data": {
                "description": "Invalid data",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "The given data was invalid."
                                },
                                "error": {
                                    "properties": {
                                        "email": {
                                            "type": "string",
                                            "example": "The email must be a valid email address."
                                        }
                                    },
                                    "type": "object"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
...
    },
...
    "security": [
        []
    ]
}